feat: ADMIN, GeneralRoleType 역할에 대한 보호로직 추가
ADMIN역할을 가진 유저가 아니면 ADMIN역할을 가진 유저에 대해 역할 변경을 할 수 없다
ADMIN역할에 대한 권한 변경은 불가능하다
ADMIN, USER 등 코드에서 직접 사용되는 GeneralRoleType에 대한 삭제는 불가능하다

+++ src/main/java/com/BaGulBaGul/BaGulBaGul/domain/user/service/RoleServiceImpl.java
@@ -6,6 +6,10 @@ import com.BaGulBaGul.BaGulBaGul.domain.user.dto.service.request.RoleRegisterReq
 import com.BaGulBaGul.BaGulBaGul.domain.user.dto.service.request.SearchRoleRequest;
 import com.BaGulBaGul.BaGulBaGul.domain.user.repository.RoleRepository;
 import com.BaGulBaGul.BaGulBaGul.domain.user.repository.querydsl.FindRoleByCondition.RoleNamesWithTotalCount;
+import com.BaGulBaGul.BaGulBaGul.global.auth.constant.GeneralRoleType;
+import com.BaGulBaGul.BaGulBaGul.global.exception.GeneralException;
+import com.BaGulBaGul.BaGulBaGul.global.response.ResponseCode;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -42,6 +46,10 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public void deleteRole(String roleName) {
+        //GeneralRoleType은 보호된다.
+        if(Arrays.stream(GeneralRoleType.values()).anyMatch(r -> r.name().equals(roleName))) {
+            throw new GeneralException(ResponseCode.FORBIDDEN);
+        }
         //연결된 권한은 cascade로 삭제
         roleRepository.deleteById(roleName);
     }
